Make Servo class templated
Users can now instantiate ServoFloat and ServoDouble in addition to Servo.
This is to allow using radians, which would get truncated with integers.
Also when using degrees this is beneficial if you want better resolution than 1 degree.
The implementation must be in the header file for template classes.
ServoBase class is shared between all template versions of ServoTemplate, so the static variable channel_next_free is shared.
Added an example for how to use ServoFloat with radians.